Improve the `convert-model.py` script for performance and footprint and add ViT models (#208)
* add 5 vit models

* free-up disk space

* bypass conversion if target already exists

+++ tools/convert-models.py
@@ -13,8 +13,8 @@ def upload_blob(bucket_name, source_file_name, destination_blob_name):
     s3 = boto3.client('s3')
 
     s3.upload_file(
-      source_file_name, 
-      bucket_name, 
+      source_file_name,
+      bucket_name,
       destination_blob_name
     )
 
@@ -23,6 +23,15 @@ def upload_blob(bucket_name, source_file_name, destination_blob_name):
             source_file_name, destination_blob_name
         )
     )
+
+def blob_exist(bucket_name, blob_name):
+    """Check if file already exists in s3 bucket"""
+    s3 = boto3.client("s3")
+    try:
+        s3.head_object(Bucket=bucket_name, Key=blob_name)
+        return True
+    except s3.exceptions.NoSuchKey:
+        return False
 
 models = {
   'alexnet': 'https://download.pytorch.org/models/alexnet-owt-4df8aa71.pth',
@@ -75,20 +84,33 @@ models = {
   'efficientnet_b7': 'https://download.pytorch.org/models/efficientnet_b7_lukemelas-c5b4e57e.pth',
   'efficientnet_v2_s': 'https://download.pytorch.org/models/efficientnet_v2_s-dd5fe13b.pth',
   'efficientnet_v2_m': 'https://download.pytorch.org/models/efficientnet_v2_m-dc08266a.pth',
-  'efficientnet_v2_l': 'https://download.pytorch.org/models/efficientnet_v2_l-59c71312.pth'
-}
+  'efficientnet_v2_l': 'https://download.pytorch.org/models/efficientnet_v2_l-59c71312.pth',
+  'vit_b_16': 'https://download.pytorch.org/models/vit_b_16-c867db91.pth',
+  'vit_b_32': 'https://download.pytorch.org/models/vit_b_32-d86f8d99.pth',
+  'vit_l_16': 'https://download.pytorch.org/models/vit_l_16-852ce7e3.pth',
+  'vit_l_32': 'https://download.pytorch.org/models/vit_l_32-c7638314.pth',
+  'vit_h_14': 'https://download.pytorch.org/models/vit_h_14_swag-80465313.pth',
+  }
 
 os.makedirs("models", exist_ok=True)
 
 for name, url in models.items():
-  m = load_state_dict_from_url(url, progress=False)
-  converted = {}
-  for nm, par in m.items():
-    converted.update([(nm, par.clone())])
   fpath = "models/" + name + ".pth"
-  torch.save(converted, fpath, _use_new_zipfile_serialization=True)
-  upload_blob(
-    "torch-pretrained-models",
-    fpath,
-    "models/vision/v2/" + fpath
-  )
+
+  if blob_exist("torch-pretrained-models", f"models/vision/v2/{fpath}"):
+    print(f"--- file {fpath} is already in the bucket. Bypassing conversion")
+
+  else:
+    # download from url, convert and upload the converted weights
+    m = load_state_dict_from_url(url, progress=False)
+    converted = {}
+    for nm, par in m.items():
+      converted.update([(nm, par.clone())])
+    torch.save(converted, fpath, _use_new_zipfile_serialization=True)
+    upload_blob(
+      "torch-pretrained-models",
+      fpath,
+      "models/vision/v2/" + fpath
+    )
+    # free disk space
+    os.remove(fpath)
